+++ js/raf.js
@@ -1,57 +1,97 @@
-const _fade = function({speed, previous, next}) {
-  if(previous === next) return;
-
-  previous.style.opacity = previous.style.opacity ? Number(previous.style.opacity) - speed : 1;
-  next.style.opacity = next.style.opacity ? Number(next.style.opacity) + speed : 0;
-
-  if(previous.style.opacity <= 0) {
-    previous.style.opacity = 0;
-    next.style.opacity = 1;
-    return;
+class r_NavigatorStyleSceneChange {
+  constructor({
+    sceneList,
+    leftButton, rightButton,
+    leftAnimation, rightAnimation,
+    navigatorWrap, navigatorButtonTemplate, activeNavigatorButtonStyle
+  }) {
+    this.sceneList = sceneList;
+    this.elLeftButton = leftButton;
+    this.elRightButton = rightButton;
+    this.leftAnimation = leftAnimation;
+    this.rightAnimation = rightAnimation;
+    this.elNavigatorWrap = navigatorWrap;
+    this.navigatorButtonTemplate = navigatorButtonTemplate;
+    this.activeStyle = activeNavigatorButtonStyle;
+    this.currentIndex = 0;
+    this.sceneLength = sceneList.length;
   }
 
-  requestAnimationFrame(() => {
-    _fade({speed, previous, next});
-  })
-};
-
-const _slide = function ({speed, type, previous, next}) {
-  if(previous === next) return;
-  
-  if(previous.style.transform === '') previous.style.transform = `translateX(0%)`;
-  if(next.style.transform === '') next.style.transform = `translateX(${type === 'left' ? -100 : 100}%)`;
-
-  let nextTranslateValue = Math.abs(next.style.transform.match(/-?\d{1,}/)[0]);
-
-  if(nextTranslateValue < 1) {
-    previous.style.transform = '';
-    next.style.transform = '';
-    previous.style.visibility = 'hidden';
-    return;
+  init() {
+    this._renderNavigatorButton({
+      sceneLength: this.sceneList.length,
+      navigatorWrap: this.elNavigatorWrap,
+      template: this.navigatorButtonTemplate
+    });
+    this._registerAllEventListener();
+    this._activateScene(this.sceneList[this.currentIndex]);
+    this._activateNavigatorButton({index: this.currentIndex, style: this.activeStyle});
   }
 
-  nextTranslateValue = nextTranslateValue * speed;
-  let previousTranslateValue = nextTranslateValue - 100;
-
-  if(type === 'left') {
-    previousTranslateValue *= -1;
-    nextTranslateValue *= -1;
+  _renderNavigatorButton({sceneLength, navigatorWrap, template}) {
+    for(let i = 0; i < sceneLength; i++) {
+      navigatorWrap.innerHTML += template(i);
+    }
   }
 
-  previous.style.transform = `translate(${previousTranslateValue}%)`;
-  next.style.transform = `translate(${nextTranslateValue}%)`;
+  _registerAllEventListener() {
+    this.elLeftButton.addEventListener('click', () => {
+      const nextIndex = this._calculateNextIndex(-1);
+      this._animateScene({previousIndex: this.currentIndex, nextIndex, animation: this.leftAnimation})
+      this._activateNavigatorButton({index: nextIndex, style: this.activeStyle});
+      this.currentIndex = nextIndex;
+    })
 
-  requestAnimationFrame(() => {
-    _slide({speed, type, previous, next});
-  })
+    this.elRightButton.addEventListener('click', () => {
+      const nextIndex = this._calculateNextIndex(1);
+      this._animateScene({previousIndex: this.currentIndex, nextIndex, animation: this.rightAnimation})
+      this._activateNavigatorButton({index: nextIndex, style: this.activeStyle});
+      this.currentIndex = nextIndex;
+    })
+
+    this.elNavigatorWrap.addEventListener('click', ({target}) => {
+      if(target.tagName !== 'LI') return;
+
+      const nextIndex = Number(target.dataset.index);
+      const animation = this.currentIndex < nextIndex ? this.rightAnimation : this.leftAnimation;
+
+      this._animateScene({previousIndex: this.currentIndex, nextIndex, animation});
+      this._activateNavigatorButton({index: nextIndex, style: this.activeStyle});
+      this.currentIndex = nextIndex;
+    })
+  }
+
+  _calculateNextIndex(number) {
+    return ((this.currentIndex + number) + this.sceneLength) % this.sceneLength;
+  }
+
+  _animateScene({previousIndex, nextIndex, animation}) {
+    const previousScene = this.sceneList[previousIndex];
+    const nextScene = this.sceneList[nextIndex];
+
+    this._activateScene(previousScene, nextScene);
+
+    animation({previous: previousScene, next: nextScene});
+  }
+
+  _activateScene(...targetSceneList) {
+    this.sceneList.forEach(scene => {
+      scene.style.visibility = 'hidden';
+    })
+
+    targetSceneList.forEach(scene => {
+      scene.style.visibility = 'visible';
+    })
+  }
+
+  _activateNavigatorButton({index, style}) {
+    const navigatorButtons = this.elNavigatorWrap.childNodes;
+    
+    navigatorButtons.forEach(button => {
+      const buttonIndex = Number(button.dataset.index);
+      buttonIndex === index ? button.classList.add(style) : button.classList.remove(style);
+    })
+  }
 }
 
-const objectCurry = function({callback, numberOfKey}) {
-  return function f(obj) {
-    if(Object.keys(obj).length >= numberOfKey) return callback(obj);
-    return (restObj) => f(Object.assign(restObj, obj));
-  }
-}
-
-export const fade = objectCurry({callback: _fade, numberOfKey: 3});
-export const slide = objectCurry({callback: _slide, numberOfKey: 4});
+export {r_NavigatorStyleSceneChange}
